tesla_rag/requirements.txt
langchain==0.0.311
google-generativeai==0.2.0
python-dotenv==1.0.0
chromadb==0.4.13
pypdf==3.16.1
beautifulsoup4==4.12.2
requests==2.31.0
wikipedia==1.4.0
tiktoken==0.5.1
pandas==2.1.1
numpy==1.26.0